+++ install-spotify.sh
@@ -1,10 +1,15 @@
-#!/bin/sh
+#!/bin/bash -e
 
-echo "Installing Spotify Connect (spotifyd)"
+SPOTIFYD_VERSION=0.2.5
+SPOTIFYD_BUILD=2019-02-25
 
-wget https://github.com/Spotifyd/spotifyd/releases/download/v0.2.5/spotifyd-2019-02-25-armv6.zip
-unzip spotifyd-2019-02-25-armv6.zip
-rm spotifyd-2019-02-25-armv6.zip
+echo -n "Do you want to install Spotify Connect (spotifyd v${SPOTIFYD_VERSION})? [y/N] "
+read REPLY
+if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then exit 0; fi
+
+wget https://github.com/Spotifyd/spotifyd/releases/download/v${SPOTIFYD_VERSION}/spotifyd-${SPOTIFYD_BUILD}-armv6.zip
+unzip spotifyd-${SPOTIFYD_BUILD}-armv6.zip
+rm spotifyd-${SPOTIFYD_BUILD}-armv6.zip
 mkdir -p /opt/local/bin
 mv spotifyd /opt/local/bin
